initial configuration
(cherry picked from commit 2e52ade223e6caaaed092b91090f060d348862cf)
