+++ src/main/java/alfio/config/authentication/support/RecaptchaLoginFilter.java
@@ -14,11 +14,10 @@
  * You should have received a copy of the GNU General Public License
  * along with alf.io.  If not, see <http://www.gnu.org/licenses/>.
  */
-package alfio.config.support.auth;
+package alfio.config.authentication.support;
 
-import alfio.manager.system.OpenIdAuthenticationManager;
-import lombok.extern.log4j.Log4j2;
-import org.springframework.security.core.context.SecurityContextHolder;
+import alfio.manager.RecaptchaService;
+import alfio.manager.system.ConfigurationManager;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 import org.springframework.security.web.util.matcher.RequestMatcher;
 import org.springframework.web.filter.GenericFilterBean;
@@ -30,28 +29,35 @@ import javax.servlet.ServletResponse;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-@Log4j2
-public class OpenIdAuthenticationFilter extends GenericFilterBean {
-    private final RequestMatcher requestMatcher;
-    private final OpenIdAuthenticationManager openIdAuthenticationManager;
 
-    public OpenIdAuthenticationFilter(String loginURL, OpenIdAuthenticationManager openIdAuthenticationManager) {
-        this.requestMatcher = new AntPathRequestMatcher(loginURL, "GET");
-        this.openIdAuthenticationManager = openIdAuthenticationManager;
+import static alfio.model.system.ConfigurationKeys.ENABLE_CAPTCHA_FOR_LOGIN;
+
+public class RecaptchaLoginFilter extends GenericFilterBean {
+    private final RequestMatcher requestMatcher;
+    private final RecaptchaService recaptchaService;
+    private final String recaptchaFailureUrl;
+    private final ConfigurationManager configurationManager;
+
+
+    public RecaptchaLoginFilter(RecaptchaService recaptchaService,
+                                String loginProcessingUrl,
+                                String recaptchaFailureUrl,
+                                ConfigurationManager configurationManager) {
+        this.requestMatcher = new AntPathRequestMatcher(loginProcessingUrl, "POST");
+        this.recaptchaService = recaptchaService;
+        this.recaptchaFailureUrl = recaptchaFailureUrl;
+        this.configurationManager = configurationManager;
     }
+
 
     @Override
     public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
         HttpServletRequest req = (HttpServletRequest) request;
         HttpServletResponse res = (HttpServletResponse) response;
-
-        if (requestMatcher.matches(req)) {
-            if (SecurityContextHolder.getContext().getAuthentication() != null || req.getParameterMap().containsKey("logout")) {
-                res.sendRedirect("/admin/");
-                return;
-            }
-            log.trace("calling buildAuthorizeUrl");
-            res.sendRedirect(openIdAuthenticationManager.buildAuthorizeUrl());
+        if (requestMatcher.matches(req) &&
+            configurationManager.getForSystem(ENABLE_CAPTCHA_FOR_LOGIN).getValueAsBooleanOrDefault() &&
+            !recaptchaService.checkRecaptcha(null, req)) {
+            res.sendRedirect(recaptchaFailureUrl);
             return;
         }
